pyproject.toml
[tool.poetry]
name = "students-python-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "models/**/*.py" },
    { include = "extensions/**/*.py" },
]

[tool.poetry.dependencies]
Flask = "^2.3.2"
Flask-SQLAlchemy = "^3.0.3"
flask-restx = "^1.1.0"
python = "^3.11"
pytest = "^7.3.1"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
